pyproject.toml
[tool.poetry]
name = "fastapi-todo-app"
version = "0.1.0"
description = "FastAPI Todo App"
authors = ["beary <bearyme@outlook.com>"]
packages = [{ include = "app" }]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.6"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
asyncpg = "^0.30.0"
pydantic-settings = "^2.7.0"
pyjwt = "^2.10.1"
alembic = "^1.14.0"
python-dotenv = "^1.0.1"
tzdata = "^2024.2"
bcrypt = "^4.2.1"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
isort = "^5.13.2"
mypy = "^1.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
